✨Feat: 메모 등록 시 date 초기값 오늘날짜로 세팅, 메모 삭제 시 사용자에게 확인 요청

+++ src/Components/ReviewMemoCard.js
@@ -18,12 +18,15 @@ const ReviewMemoCard = ({
   const user = localStorage.getItem("anonUserid");
 
   const handleDelete = async () => {
-    try {
-      await deleteMemo(user, recordId, pinId, memoId);
-      onDeleteSuccess(); // 부모 컴포넌트에 삭제 성공 알림
-      console.log("Memo deleted successfully");
-    } catch (error) {
-      console.error("Error deleting memo: ", error);
+    const isConfirmed = window.confirm("이 메모를 삭제할까요?");
+    if (!isConfirmed) {
+      try {
+        await deleteMemo(user, recordId, pinId, memoId);
+        onDeleteSuccess(); // 부모 컴포넌트에 삭제 성공 알림
+        console.log("Memo deleted successfully");
+      } catch (error) {
+        console.error("Error deleting memo: ", error);
+      }
     }
   };
